权限判断

+++ app/Http/Controllers/API/Admin/AcceptMessagesController.php
@@ -10,6 +10,7 @@ use App\Models\Employee;
 use App\Repositories\AcceptMessagesRepository;
 use App\Services\MessagesService;
 
+
 class AcceptMessagesController extends APIBaseController
 {
     protected $repo;
@@ -56,6 +57,9 @@ class AcceptMessagesController extends APIBaseController
     //消息接收人员列表
     public function index(MessagesService $service)
     {
+        if (empty(Common::user()->can('accept_message_list'))) {
+            return $this->sendError('无accept_message_list权限','403');
+        }
         $res= $this->repo->messageList($this->req, $service);
         return $this->sendResponse($res,'获取成功');
     }
@@ -63,6 +67,9 @@ class AcceptMessagesController extends APIBaseController
     //添加消息接收人员
     public function store(MessagesService $service)
     {
+        if (empty(Common::user()->can('add_accept_message'))) {
+            return $this->sendError('无add_accept_message权限','403');
+        }
         $res = $this->repo->addAcceptMessage($this->req, $service);
         return $this->sendResponse($res, '设置成功');
     }
@@ -70,6 +77,9 @@ class AcceptMessagesController extends APIBaseController
     //删除消息接收人员
     public function destroy()
     {
+        if (empty(Common::user()->can('del_accept_message'))) {
+            return $this->sendError('无del_accept_message权限','403');
+        }
         $res = AcceptMessage::where(['type' => $this->req->type,'employee_id' => $this->req->employee_id])->delete();
         return $this->sendResponse($res,'删除成功');
     }
