+++ app/Http/Controllers/API/Admin/InformationController.php
@@ -7,6 +7,7 @@ use App\Models\Information;
 use App\Repositories\InformationRepository;
 use App\Services\InformationService;
 use Illuminate\Http\Request;
+use App\Handler\Common;
 
 class InformationController extends APIBaseController
 {
@@ -14,6 +15,9 @@ class InformationController extends APIBaseController
         InformationRepository $repository
     )
     {
+        if (empty(Common::user()->can('information_list'))) {
+            return $this->sendError('无information_list权限','403');
+        }
         $res = $repository->informationList();
         return $this->sendResponse($res, '获取资讯列表数据成功');
     }
@@ -23,6 +27,9 @@ class InformationController extends APIBaseController
         InformationRepository $repository
     )
     {
+        if (empty(Common::user()->can('add_information'))) {
+            return $this->sendError('无add_information权限','403');
+        }
         $res = $repository->addInformation($request);
         return $this->sendResponse($res,'添加资讯成功');
     }
@@ -48,6 +55,9 @@ class InformationController extends APIBaseController
         Information $information
     )
     {
+        if (empty(Common::user()->can('del_information'))) {
+            return $this->sendError('无del_information权限','403');
+        }
         $res = $information->delete();
         return $this->sendResponse($res,'资讯删除成功');
     }
@@ -57,6 +67,9 @@ class InformationController extends APIBaseController
         InformationRepository $repository
     )
     {
+        if (empty(Common::user()->can('set_top'))) {
+            return $this->sendError('无set_top权限','403');
+        }
         $res = $repository->setTop($id);
         return $this->sendResponse($res, '置顶成功');
     }
@@ -66,6 +79,9 @@ class InformationController extends APIBaseController
         InformationRepository $repository
     )
     {
+        if (empty(Common::user()->can('del_top'))) {
+            return $this->sendError('无del_top权限','403');
+        }
         $res = $repository->delTop($id);
         return $this->sendResponse($res, '取消置顶成功');
     }
