+++ app/Http/Controllers/API/Admin/PcRecommendsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\API\APIBaseController;
 use App\Http\Requests\Admin\PcRecommendsRequest;
 use App\Models\PcRecommend;
 use App\Repositories\PcRecommendsRepository;
+use App\Handler\Common;
 
 class PcRecommendsController extends APIBaseController
 {
@@ -14,6 +15,9 @@ class PcRecommendsController extends APIBaseController
         PcRecommendsRepository $repository
     )
     {
+        if (empty(Common::user()->can('Pc_recommends_list'))) {
+            return $this->sendError('无Pc_recommends_list权限','403');
+        }
         $res = $repository->pcRecommendsList();
         return $this->sendResponse($res,'pc端精品推荐列表获取成功');
     }
@@ -24,6 +28,9 @@ class PcRecommendsController extends APIBaseController
         PcRecommendsRepository $repository
     )
     {
+        if (empty(Common::user()->can('add_pc_recommends'))) {
+            return $this->sendError('无add_pc_recommends权限','403');
+        }
         $res = $repository->addPcRecommends($request);
         return $this->sendResponse($res,'添加pc端精品推荐成功');
     }
@@ -43,6 +50,9 @@ class PcRecommendsController extends APIBaseController
         PcRecommendsRepository $repository
     )
     {
+        if (empty(Common::user()->can('update_pc_recommends'))) {
+            return $this->sendError('无update_pc_recommends权限','403');
+        }
         $res = $repository->updatePcRecommends($request, $pcRecommend);
         if (empty($res)) return $this->sendError('修改pc端精品推荐信息失败');
         return $this->sendResponse($res,'修改pc端精品推荐信息成功');
@@ -53,6 +63,9 @@ class PcRecommendsController extends APIBaseController
         PcRecommend $pcRecommend
     )
     {
+        if (empty(Common::user()->can('del_pc_recommends'))) {
+            return $this->sendError('无del_pc_recommends权限','403');
+        }
         if (empty($res = $pcRecommend->delete())) return $this->sendError('删除pc端精品推荐信息失败');
         return $this->sendResponse($res,'删除pc端精品推荐信息成功');
     }
